# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"allergiesInterface": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"medicationsInterface": {
				"properties": {
					"name": {
						"type": "string"
					},
					"dose": {
						"type": "string"
					},
					"reason": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"dose",
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"emergencyContactInterface": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"phone",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "ie6r_apis",
		"version": "1.0.0",
		"description": "IE6R apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "STS"
		}
	},
	"paths": {
		"/user/allergy/add/update": {
			"put": {
				"operationId": "AddOrUpdateAllergy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update User Allergy",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"allergies": {
										"items": {
											"$ref": "#/components/schemas/allergiesInterface"
										},
										"type": "array"
									}
								},
								"required": [
									"allergies"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/medication/add/update": {
			"put": {
				"operationId": "AddOrUpdateMedication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update User Medication",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"medications": {
										"items": {
											"$ref": "#/components/schemas/medicationsInterface"
										},
										"type": "array"
									}
								},
								"required": [
									"medications"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/contact/add/update": {
			"put": {
				"operationId": "AddOrUpdateEmergencyContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update User Emergency Contact",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"contacts": {
										"items": {
											"$ref": "#/components/schemas/emergencyContactInterface"
										},
										"type": "array"
									}
								},
								"required": [
									"contacts"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/info/list": {
			"get": {
				"operationId": "UserInfoList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "List User Info Medics And other\ntype == 'allergy', 'medication', 'contacts'",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort_column",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort_direction",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search_key",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/info/details": {
			"get": {
				"operationId": "UserInfoDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Details User Info Medics And other\ntype == 'allergy', 'medication', 'contacts'",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "item_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/info/screen": {
			"get": {
				"operationId": "UserInfoMainScreen",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "User Info Screen Api Medics And other",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/user/info/delete": {
			"delete": {
				"operationId": "DeleteUserInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "user info delete\ntype == 'allergy', 'medication', 'contacts'",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"type": {
										"type": "string"
									},
									"item_id": {
										"type": "string"
									}
								},
								"required": [
									"type",
									"item_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/setting/update": {
			"put": {
				"operationId": "UpdateSettings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update User Settings",
				"tags": [
					"User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"dark_mode": {
										"type": "boolean"
									},
									"allow_multi_contact": {
										"type": "boolean"
									},
									"allow_gps": {
										"type": "boolean"
									},
									"send_sms": {
										"type": "boolean"
									},
									"display_dob": {
										"type": "boolean"
									}
								},
								"required": [
									"dark_mode",
									"allow_multi_contact",
									"allow_gps",
									"send_sms",
									"display_dob"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}